luffyapi/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-ckeditor==5.8.0
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.6
MarkupPy==1.14
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
pycryptodomex==3.7.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
